feature start: Fix KeyError for unestimated stories
Issue:
------
On feature start, a list of available stories from Current and Backlog is show
and the user is asked to choose. If he, however, chooses and unestimated story,
a KeyError is fired.

Solution:
---------
Do not show any unestimated story in the dialog in the first place. Then,
logically, the user cannot choose it to get the KeyError.