Ask user to confirm the release
Up until now, release start would simply start the release, failing when
the branch for the release was already existing, not even pushing the
branch upstream after a successful creation.

Now, the user is being asked if he wants to start the release after
seeing all the stories that will be a part of it. release start also
works for updating releases as well. It adds new stories to the release
every time it is called. It also does not fail if the release branch
exists already, it simply aborts. Last but not least, when a release
branch is created, it is not pushed upstream automatically.