feature start: Enforce a single branch per story
Issue
-----
Up until now it was possible to choose a story in the feature start
dialog and then choose arbitrary description to form the full branch
name. This, however, has the flaw that it was possible to create
multiple branches for a single Pivotal Tracker story. This is not
supposed to be possible since branches are often identified by their
associated story id.

Solution
--------
During the feature start dialog, a check is performed to see if there is
not an existing branch already associated to that particular story id.
If there is, the user is given choice to checkout that already existing
branch or not.