init: Let user pick up his PT project and RB repo
Motivation
----------
Up until now it was necessary to issue git config to manually assign a
Pivotal Tracker project and a Review Board repository to the cloned git
repository for the gitflow plugin to work properly, to recognize all the
services and to be able to connect to them. That was, however, extremely
annoying for the users, to go to visit all the tools and manually get
the IDs every time they clone a repository.

Solution
--------
This changeset introduces a dialog, which is shown during flow init. The
user is given choice between all the projects he can access. After
picking up a PT project and a RB repository, their IDs are automatically
saved into .git/config.