release finish: Make sure QA label is checked
Issue
-----
On release finish, qa+ label was not checked so all the stories were
released even though they were not marked as QA'd.

Solution
--------
Iterate over all the stories and look for qa+ label before doing
anything else.

release finish was also a bit refactored to perform the release steps in
a different order so that it makes more sense, i.e. the more important
operations and checks are performed first.