feature start: Handle PT bug stories correctly
Issue
-----
Pivotal Tracker part of the codebase was containing asserts in the Story class
to make sure that some methods are called only on feature stories. However, the
same methods were supposed to work for bugs as well and there the asserts were
incorrectly failing.

Solution
--------
Add 'or is_bug()' into asserts, make sure that the conditionals in prompting
user for new feature are correct as well.