feature finish: Generate diff for the whole branch
The issue with the previous implementation was that when someone got a
review for his or her branch and posted an update, the update containes
only the changes since the previous patch set. This is, however, not how
Review Board works. You need to post the full diff every time you want
to create/update a review request.

Change-Id: Ie91be3c913050115476e5ba0537b28ed5187f684
Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>
Reviewed-on: http://dev.salsitasoft.com/gerrit/2
Reviewed-by: Tomas Brambora <tomas@salsitasoft.com>
Tested-by: Ondrej Kupka <ondrejk@salsitasoft.com>