feature finish: Post comment into Pivotal Tracker
Once code review is posted into Review Board, a comment is added into
Pivotal Tracker. This comment contains the link to the review request
page for easy navigation.

One caveat is, though, that this link is sent after a review request
draft is posted, so theoretically someone can follow the link before the
review request is actually published. But this is very improbable to happen.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>
Change-Id: I5b099e3b16042407ebc8f62a1e47e584d107e84b