feature finish: Choose merge branch correctly
feature finish was somehow depending on the state of the story to decide
what to merge, if into release or into develop, namely it was not
merging into release if the story was not finished. This, however,
breaks if the story was restarted. Then it is assigned to a release but
not finished.

This patch makes the plugin decide on the merge branch solely on the
Pivotal Tracker story label. If the release label is present, the story
is merged into release, otherwise it's develop.

Change-Id: Iba4fe7055e16d1a2a69a20086ba77509573d1a52
Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>