release finish: Make it more idempotent
The problem is that once a review is submitted, Review Boards stops to
return it in the call to its API unless you specify status='all'. So if
you call release finish multiple times, it fails because it cannot find
the reviews submitted during the previous call.

This patch makes gitflow add status='all' when calling the API and
detects better what the status of the branch review is.

Change-Id: Id78d5555a838176f1ce5cb4d14b45ce9583ee015
Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>