feature finish: Do not add links into merge msg
The problem is that the merge commit needs amending because of the order of
steps that is given. If you, however, need to resolve a merge conflict and push
manually, this amend step inserting the links will fail, because the merge
commit was already pushed upstream manually, which is not what the plugin counts
on. It would be necessary to use git push -f to make amend work, but that is too
dangerous. We decided to simply not append any links for now.

Change-Id: Ie2781dd2d56d509ffcf20d94a0157617a8c2ce3b
Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>