feature finish: Be sure to fetch before merging
The issue was that must_be_uptodate was issuing git fetch origin develop, but
this command does not update remote-tracking branches. The result was that the
command was failing on push instead of merge.

This patch makes sure that we really fail as soon as possible, that is before
even trying to merge.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>