release finish: Check reviews before submitting
The issue was that during release finish, the relevant stories were iterated
over and their reviews were submitted one by one. This create an inconsistent
situation once any of the stories is not submitted.

This fix makes the plugin first check for all reviews in one step, and then, if
everything is ok, submit them all at once so that there is no space for any
inconsistency.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>