feature start: Handle rejected stories properly
Up until now it was not clear how rejected stories are being handled.

This patch adds explicit handling of rejected stories. They are offered during
feature start, but then it just performs checkout and PT story start. The the
branch does not exist, it prints a warning and fails since the story must exist
because someone must have called feature start in the past, and thus create the
feature branch.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>