feature finish: Make rbt error output available
Using subprocess.check_output for the call to rbt was not the best decision
since it just throws an exception in case the return code is not 0. In other
works, there is no way we can get out hand on the error output.

This patch makes the plugin call rbt using subprocess.Popen, using pipes to get
standard error output and print it to the user in case the return code is not 0.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>