release finish: Submit reviews after merging ok
There was a single step to check and then submit all the code reviews,
and the another step to merge branches. But what if the reviews are ok,
they are submitted, and then we fail to merge?

That single Review Board step is now split, so we

    1) Check that all relevant review requests are there.
    2) Merge.
    3) Submit all relevant review requests.

We cannot get into an inconsistent state with this sequence of steps
unless we get an internal server error from Review Board.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>