feature start & finish: Use marker to get diff
Since there were some issues with getting the feature base correctly, it was
decided that we would use a marker branch to mark the base. So from now on,
there is another branch created for each feature branch called
'base_feature/<story_id>' and it is being used as the base for computing
feature diff.

This marker branch is pushed on feature publish as well.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>