Create branch automatically if remote one exists
When there is origin/release/0.0.1 and you are finishing a story
assigned to that release, it will fail with release/0.0.1 branch not
found. But it exists in the origin!

From now on, if you try to look for a branch, origin is always checked
as well and if a branch of the same name exists, its local counterpart
is created and set to track the remote branch.