init: Rewrite ssh:// origin remote format
If you clone using ssh://, rbt will fail to post code review request
since that is not the URL Review Board uses to identify the repository.
It uses the SSH format.

This patch teaches flow init to automatically rewrite ssh:// to the SSH
format so that the error thrown later is avoided.