release finish: Do not try to redeliver stories
The plugin always tries to deliver all the stories during release
finish. That counts on the fact that they are all in the finished state.
But that does not have to be the case if they are already accepted or
delivered.

So from now on the plugin is skipping stories during delivery if these
stories are already delivered or accepted/rejected.