Add support for multiple repositories per project
From now on it is possible to use multiple repositories for a single PT project.
This is implemented using PT labels, which are then used during release
operations to pick up the stories associated with the repository you are
currectly working with.

This is implemented by asking the user for a PT label during flow init, and
later on filtering PT stories according to this label. The label must start with
'repo-' so that GF can distinguish it from any other random PT label.