multiple repos: Implement include/exclude
From now on it is not necessary to use repo- prefix from the repository labels.
Instead of it you define a label filter for each repository. The filter either
includes a single PT label, or excludes any number of PT labels.

This is again being asked during git flow init. If no include label is defined,
the exclude labels question is asked and it's expecting a comman-separated list
of labels. If both questions are answered with an empty reply, gitflow behaves
exactly as it used to.